+++ src/main/java/at/hannibal2/skyhanni/events/entity/EntityAttributeUpdateEvent.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.events
+package at.hannibal2.skyhanni.events.entity
 
 import at.hannibal2.skyhanni.api.event.GenericSkyHanniEvent
 import at.hannibal2.skyhanni.skyhannimodule.PrimaryFunction
@@ -6,8 +6,8 @@ import net.minecraft.core.Holder
 import net.minecraft.world.entity.LivingEntity
 import net.minecraft.world.entity.ai.attributes.Attribute
 
-@PrimaryFunction("onAttributeWatcherUpdate")
-data class AttributeWatcherUpdateEvent<T : LivingEntity>(
+@PrimaryFunction("onEntityAttributeUpdate")
+data class EntityAttributeUpdateEvent<T : LivingEntity>(
     val entity: T,
     val attribute: Holder<Attribute>,
 ) : GenericSkyHanniEvent<T>(entity.javaClass)
